finishing touches, so the README and broke the code into components
